pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nemo-guardrails-cai"
version = "0.1.0"
description = "NVIDIA NeMo Guardrails deployment for Cloudera AI"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Cloudera AI Team"}
]
keywords = [
    "nemo",
    "guardrails",
    "llm",
    "safety",
    "cloudera",
    "ai",
    "machine-learning",
    "conversational-ai"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "nemoguardrails>=0.9.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
cai = [
    "caikit>=0.1.0",
]
local-models = [
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "sentencepiece>=0.1.99",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "httpx>=0.26.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]
all = [
    "caikit>=0.1.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "sentencepiece>=0.1.99",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "httpx>=0.26.0",
]

[project.urls]
Homepage = "https://github.com/cloudera/nemo-guardrails-cai"
Documentation = "https://github.com/cloudera/nemo-guardrails-cai/blob/main/README.md"
Repository = "https://github.com/cloudera/nemo-guardrails-cai"
"Bug Tracker" = "https://github.com/cloudera/nemo-guardrails-cai/issues"

[project.scripts]
nemo-guardrails-server = "nemo_guardrails_cai.server:main"
nemo-guardrails-model-service = "model_service:main"

[tool.setuptools.packages.find]
include = ["nemo_guardrails_cai*"]
exclude = ["tests*", "docs*", "examples*"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=nemo_guardrails_cai --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

# MyPy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "nemoguardrails.*",
]
ignore_missing_imports = true

# Coverage configuration
[tool.coverage.run]
source = ["nemo_guardrails_cai"]
omit = [
    "*/tests/*",
    "*/examples/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
